+++ game.py
@@ -15,954 +15,17 @@ import pickle5.pickle as pickle
 import pygame 
 from pygame.locals import *
 
-actions = {
-	"up":     pygame.K_z,
-	"down":   pygame.K_s,
-	"right":  pygame.K_d,
-	"left":   pygame.K_q,
-	"sprint": pygame.K_LSHIFT,
-	"attack": pygame.K_SPACE,
-	"inventory": pygame.K_i,
-	"skillTree": pygame.K_TAB,
-	"pause": pygame.K_ESCAPE
-}
-
-weaponTypeAPSMultiplier = {
-	'sword': 1,
-	'mace': 2
-}
-
-class HUDMenuManager(gregngine.HUDMenuManager):
-	def __init__(self,engine,hudScale):
-		super().__init__(engine,hudScale)
-
-		self.param = {
-			'margin': 5,
-			'fontSize': 20,
-			'padding': {
-				'height': 6,
-				'width': 10
-			},
-			'radius': 10,
-			'alpha': {
-				'normal': 50,
-				'hover': 150
-			},
-			'color': (255,255,255,255)
-		}
-
-		self.font = pygame.font.Font('./assets/fonts/Pixel Digivolve.otf', self.param['fontSize']*self.hudScale)
-
-		self.huds = {
-			"startMenu": {
-				'fonction': self.startMenu,
-				'buttons': [
-					{
-						'text': 'Jouer',
-						'fonction': self.btnPlay
-					},
-					{
-						'text': 'Options',
-						'fonction': self.btnOptions
-					},
-					{
-						'text': 'Quitter',
-						'fonction': self.btnQuit
-					}
-				]
-			},
-			"pauseMenu": {
-				'fonction': self.pauseMenu,
-				'isOpen': False,
-				'buttons': [
-					{
-						'text': 'retour',
-						'fonction': self.btnResume
-					},
-					{
-						'text': 'sauvegarder',
-						'fonction': self.btnSave
-					},
-					{
-						'text': 'options',
-						'fonction': self.btnOptions
-					},
-					{
-						'text': 'quitter',
-						'fonction': self.btnQuit
-					}
-				]
-			}
-		}
-	
-	def startMenu(self, passThrough):
-		screenWidth, screenHeight = self.engine.HUDSurface.get_size()
-		self.engine.window.fill((255,255,255))
-
-		scale = self.hudScale
-		
-		imgWidth, imgHeight = self.font.size('h')
-		imgHeight += self.param['padding']['height']*2*scale
-		topHeight = len(self.huds["pauseMenu"]['buttons'])*(imgHeight+self.param['margin']*scale)
-		top = int(screenHeight - topHeight)
-
-		self.mousePos = passThrough['mousePos']
-
-		btns = pygame.Surface((screenWidth, screenHeight), pygame.SRCALPHA)
-
-		for button in self.huds["startMenu"]['buttons']:
-			height = self.drawButton(button,'centered',top,passThrough,btns)
-			top += height*scale
-
-		title = self.font.render('The Legend of Pylda', False, (255, 158, 54))
-		imgLeft = int(self.engine.window.get_width()/2 - title.get_width()/2) + scale
-		imgTop = 20*screenHeight/100
-		self.engine.window.blit(title, (imgLeft,imgTop))
-
-		self.engine.window.blit(btns, self.engine.renderedDisplayOffsets)
-	
-	def pauseMenu(self, passThrough):
-		screenWidth, screenHeight = self.engine.HUDSurface.get_size()
-
-		background = pygame.Surface((screenWidth, screenHeight), pygame.SRCALPHA)
-		background.fill((0,0,0,75))
-		self.engine.window.blit(background, self.engine.renderedDisplayOffsets)
-
-		scale = self.hudScale
-		
-		imgWidth, imgHeight = self.font.size('h')
-		imgHeight += self.param['padding']['height']*2*scale
-		topHeight = len(self.huds["pauseMenu"]['buttons'])*(imgHeight+self.param['margin']*scale)
-		top = int(screenHeight/2 - topHeight/2)
-
-		self.mousePos = passThrough['mousePos']
-
-		btns = pygame.Surface((screenWidth, screenHeight), pygame.SRCALPHA)
-
-		for button in self.huds["pauseMenu"]['buttons']:
-			height = self.drawButton(button,'centered',top,passThrough,btns)
-			top += height*scale
-
-		self.engine.window.blit(btns,  self.engine.renderedDisplayOffsets)
-
-	def drawButton(self,button,rectLeft,rectTop,passThrough,btns):
-		scale = self.hudScale
-
-		color = button['color'] if 'color' in button else self.param['color']
-		img = self.font.render(button['text'], False, color)
-
-		rectHeight = img.get_height() + self.param['padding']['height']*2*scale
-		rectWidth = img.get_width() + self.param['padding']['width']*2*scale
-
-		imgLeft = int(passThrough['window'].get_width()/2 - img.get_width()/2) + scale
-		imgTop = int(rectTop + rectHeight/2  - img.get_height()/2) - scale
-
-		if rectLeft == 'centered':
-			rectLeft = int(passThrough['window'].get_width()/2 - rectWidth/2)
-
-		collisionRect = Rect((rectLeft,rectTop),(int(rectWidth),int(rectHeight)))
-		
-		isHover = collisionRect.collidepoint(self.mousePos)
-		alpha = self.param['alpha']['hover'] if isHover else self.param['alpha']['normal']
-
-		rects, circles = gFunction.createRectWithRoundCorner(rectLeft,rectTop,rectWidth,rectHeight,self.param['radius'])
-		for rect in rects:
-			pygame.draw.rect(btns,(0,0,0,alpha),Rect(rect))
-		for circle in circles:
-			pygame.draw.circle(btns,(0,0,0,alpha),circle,self.param['radius'])
-		
-		btns.blit(img,(imgLeft, imgTop))
-
-		# handle click requests
-		if passThrough['mousePosClick'] is not None:
-			if collisionRect.collidepoint(passThrough['mousePosClick']):
-				button['fonction']()
-		return int(rectHeight/2 + self.param['margin'])
-	
-	def btnResume(self):
-		self.huds['pauseMenu']['isOpen'] = False
-		self.engine.currentHUD = 'main'
-
-	def btnSave(self):
-		with open(self.engine.savePath + '/data.save', 'wb') as f:
-			data = {}
-			
-			data['playerPos'] = (self.engine.player.x,self.engine.player.y)
-			data['playerOrientation'] = self.engine.player.animator.lastAnimation
-
-			data['playerInv'] = {}
-			data['playerInv']['hands'] = None if self.engine.player.inventory.hands is None else self.engine.player.inventory.hands.data['name']
-			data['playerInv']['handsStack'] = self.engine.player.inventory.handsStack
-			data['playerInv']['slots'] = []
-			for row in self.engine.player.inventory.slots:
-				r = []
-				for col in row:
-					if col is not None:
-						r.append(col.data['name'])
-					else:
-						r.append(None)
-				data['playerInv']['slots'].append(r)
-
-			data['playerInv']['stacks'] = self.engine.player.inventory.stacks
-			data['playerData'] = self.engine.player.data
-			data['playerData']['skillTree']['skills'] = self.engine.player.skillTree.skills
-
-			data['entities'] = []
-			for entitie in self.engine.entitiesManager.entities:
-				dico = {}
-				if 'entityRepr' in entitie.param and entitie.param['entityRepr'] == 'player':
-					continue 
-
-				if 'entityRepr' in entitie.param and entitie.param['entityRepr'] != 'player':
-					dico['type'] = 'entitie'
-					dico['repr'] = entitie.param['entityRepr']
-					dico['stats'] = entitie.stats
-
-				elif 'itemRepr' in entitie.param:
-					dico['type'] = 'item'
-					dico['repr'] = entitie.param['itemRepr']
-
-				dico['x'], dico['y'] = entitie.x, entitie.y
-				data['entities'].append(dico)
-
-			pickle.dump(data, f, pickle.HIGHEST_PROTOCOL)
-
-		print('Save')
-
-	def btnOptions(self):
-		print('Options')
-
-	def btnQuit(self):
-		pygame.quit()
-		quit()
-
-	def btnPlay(self):
-		self.engine.currentHUD = 'main'
-		self.engine.loadGame()
-
-class Item():
-	def __init__(self,param):
-		"""
-		param = {
-			"itemRepr" : itemRepr,
-			"engine" : engine obj
-		}
-		"""
-	
-		self.param = param
-
-		self.engine = param['engine']
-		self.param["pixelSize"] = self.engine.param["pixelSize"]
-
-		if not 'x' in param:
-			self.isGrounded = False
-			self.x = 0
-			self.y = 0
-		else:
-			self.isGrounded = param['isGrounded']
-			self.x = param['x']
-			self.y = param['y']
-
-
-		with open("items/" + param['itemRepr'] + ".json","r") as file:
-			self.data = json.load(file)
-
-		self.name = self.data['displayedName']
-		
-		tilemap = self.data['tilemap']
-		self.tilemap = pygame.image.load('assets/sprites/' + self.data["tilemap"]["src"])
-		size = self.tilemap.get_size()
-
-		self.tilemapSettings = {
-			'start': ((size[0]/tilemap["totalColumns"])*tilemap["colStart"],(size[1]/tilemap["totalRow"])*tilemap["rowStart"]),
-			'size': (size[0]/tilemap["totalColumns"],size[1]/tilemap["totalRow"]),
-		}
-		
-		location = (self.tilemapSettings['start'][0], self.tilemapSettings['start'][1])
-		self.sprite = self.tilemap.subsurface(pygame.Rect(location, self.tilemapSettings['size']))
-
-		itemScale = (int(self.param['pixelSize']*tilemap['scaling']/tilemap['ratio']),int(self.param['pixelSize']*tilemap['scaling']))
-		self.groundSprite = pygame.transform.scale(self.sprite,itemScale)
-
-	def drawSprite(self,scale,coords,passThrough):
-		tilemap = self.data['tilemap']
-		itemScale = (int(coords['width']*scale*tilemap['scaling']/tilemap['ratio']),int(coords['height']*scale*tilemap['scaling']))
-		sprite = pygame.transform.scale(self.sprite.convert_alpha(),itemScale)
-		if self.data['itemType'] == 'weapon':
-			sprite = pygame.transform.rotate(sprite,-45)
-
-		self.engine.HUDSurface.blit(sprite, (coords['left']+(tilemap['leftOffset']*scale) , coords['top']+(tilemap['topOffset']*scale)))
-
-	def draw(self, xStart, yStart, passThrough):
-		if self.isGrounded:
-			xToDraw = self.x - xStart 
-			yToDraw = self.y - yStart 
-
-			self.engine.renderedSurface.blit(self.sprite , (xToDraw*self.param['pixelSize'], yToDraw*self.param['pixelSize'] - self.groundSprite.get_height()/4))
-
-	def move(self, dtime):
-		pass
-
-	def drawHud(self, xPos, yPos, passThrough):
-		pass
-
-	def death(self):
-		pass
-
-class Inventory():
-	def __init__(self,player):
-		self.slots  = [[None,None,None,None],[None,None,None,None]]
-		self.stacks = [[   0,   0,   0,   0],[   0,   0,   0,   0]]
-		self.hands = None
-		self.handsStack = 0
-		self.isOpen = False
-		self.player = player
-		self.engine = self.player.engine
-
-		self.font = pygame.font.Font('./assets/fonts/Pixel Digivolve.otf', 10*self.player.engine.param['HudScale'])
-	
-	def drawHud(self,passThrough):
-		scale = passThrough['HudScale']
-		width, height = self.engine.HUDSurface.get_size()
-		coords = {
-			'right': 8,
-			'top': 8,
-			'height': 50,
-			'width': 50,
-			'inventory-margin': 5
-		}
-		coords['left'] = width - coords['width']*scale - coords['right']*scale
-		coords['top'] *= scale
-		transparency = 150
-		radius = 10
-		radius *= scale
-
-		rects, circles = gFunction.createRectWithRoundCorner(0,0,coords['width']*scale,coords['height']*scale,radius)
-		case = pygame.Surface((coords['width']*scale,coords['height']*scale), pygame.SRCALPHA)
-		for rect in rects:
-			pygame.draw.rect(case,(0,0,0,transparency),Rect(rect))
-		for circle in circles:
-			pygame.draw.circle(case,(0,0,0,transparency),circle,radius)
-
-		if passThrough['currentHUD'] == 'main':
-			
-			self.engine.HUDSurface.blit(case, (coords['left'], coords['top']))
-
-			if self.hands is not None:
-				self.hands.drawSprite(scale,coords, passThrough)
-				#passThrough['window'].blit(sprite, (width - coords['width']*scale - coords['right']*scale -8, -26+coords['top']*scale))
-			
-			if self.handsStack > 1:
-				img = self.font.render(str(self.handsStack), False, (255, 255, 255))
-				imgLeft, imgTop = coords['left'] + case.get_width() - img.get_width() - img.get_width()/2, coords['top'] + case.get_height() - img.get_height() - img.get_height()/4
-				self.engine.HUDSurface.blit(img,(imgLeft, imgTop))
-
-			if passThrough['mousePosClick'] is not None and Rect((coords['left'], coords['top']),(int(coords['width']*scale),int(coords['height']*scale))).collidepoint(passThrough['mousePosClick']):
-				self.isOpen = True
-		
-		if passThrough['currentHUD'] == 'inventory':
-			stepHeight = coords['height'] + coords['inventory-margin']*2
-			stepWidth  = coords['width']  + coords['inventory-margin']*2
-
-			invWidth  = len(self.slots[0]) * stepWidth *scale
-			invHeight = len(self.slots)    * stepHeight *scale
-			startLeft = width/2  - invWidth/2  
-			startTop  = height/2 - invHeight/2 
-
-			cases = []
-
-			for y,y_value in enumerate(self.slots):
-				for x,x_value in enumerate(y_value):
-					placeLeft = startLeft + x*(stepWidth  + coords['inventory-margin']/2)*scale + coords['inventory-margin']*scale
-					placeTop  = startTop  + y*(stepHeight + coords['inventory-margin']/2)*scale + coords['inventory-margin']*scale
-					cases.append({'place':(int(placeLeft),int(placeTop)),'x_value':x_value,'x':x,'y':y,'ground':False})
-
-			cases.append({
-				'place':(int(width/2  - stepWidth*scale/2 + coords['inventory-margin']*1.68*scale),int(height/2  - stepHeight*scale*2 + coords['inventory-margin']*scale)),
-				'x_value':self.hands,
-				'ground':False})
-			
-			ents = [ent for ent in passThrough['visibleEntities'] if ent.data['type'] == 'item' and ent.x == round(self.player.x) and ent.y == round(self.player.y)]
-
-			if len(ents) > 0:
-				groundItemWidth  = len(ents) * stepWidth * scale
-				startLeft = width/2  - groundItemWidth/2
-				startTop  = height/2 + invHeight/2 
-
-				isSolo = 1 if len(ents) == 1 else 0
-
-				for x,entity in enumerate(ents):
-					placeLeft = startLeft + x*(stepWidth  + coords['inventory-margin']/2)*scale + coords['inventory-margin']*1.5*scale + coords['inventory-margin']/2*scale*isSolo
-					placeTop  = startTop  + coords['inventory-margin']*2*scale 
-					cases.append({'place':(int(placeLeft),int(placeTop)),'x_value':entity,'ground':True})
-
-			for caseInfo in cases:
-				self.engine.HUDSurface.blit(case,caseInfo['place'])
-				
-				if caseInfo['x_value'] is not None:
-					coords['left'],coords['top'] = caseInfo['place']
-					caseInfo['x_value'].drawSprite(scale,coords,passThrough)
-					img = self.font.render(caseInfo['x_value'].data['displayedName'], False, (255, 255, 255))
-					imgLeft, imgTop = coords['left'] + (case.get_width() - img.get_width())/2, coords['top'] + case.get_height() - scale
-					self.engine.HUDSurface.blit(img,(imgLeft, imgTop))
-
-					if 'y' in caseInfo:
-						if self.stacks[caseInfo['y']][caseInfo['x']] > 1:
-							img = self.font.render(str(self.stacks[caseInfo['y']][caseInfo['x']]), False, (255, 255, 255))
-							imgLeft, imgTop = coords['left'] + case.get_width() - img.get_width() - img.get_width()/2, coords['top'] + case.get_height() - img.get_height() - img.get_height()/4
-							self.engine.HUDSurface.blit(img,(imgLeft, imgTop))
-					elif caseInfo['x_value'] == self.hands:
-						if self.handsStack > 1:
-							img = self.font.render(str(self.handsStack), False, (255, 255, 255))
-							imgLeft, imgTop = coords['left'] + case.get_width() - img.get_width() - img.get_width()/2, coords['top'] + case.get_height() - img.get_height() - img.get_height()/4
-							self.engine.HUDSurface.blit(img,(imgLeft, imgTop))
-
-				if passThrough['mousePosClick'] is not None:
-					if Rect(caseInfo['place'],(int(coords['width']*scale),int(coords['height']*scale))).collidepoint(passThrough['mousePosClick']):
-						if 'y' in caseInfo:
-							temp = self.hands
-							stackTemp = self.handsStack
-
-							self.hands = self.slots[caseInfo['y']][caseInfo['x']]
-							self.slots[caseInfo['y']][caseInfo['x']] = temp
-
-							self.handsStack = self.stacks[caseInfo['y']][caseInfo['x']]
-							self.stacks[caseInfo['y']][caseInfo['x']] = stackTemp
-
-						elif caseInfo['ground']:
-							freeCases = []
-							sameCases = []
-							if self.hands is None:
-								freeCases.append('hands')
-							elif self.hands.data['name'] == caseInfo['x_value'].data['name'] and self.hands.data['stats']['stackable'] and self.stacks[y][x] < 100:
-								sameCases.append('hands')
-							for y,y_value in enumerate(self.slots):
-								for x,x_value in enumerate(y_value):
-									if x_value is not None and x_value.data['name'] == caseInfo['x_value'].data['name'] and x_value.data['stats']['stackable'] and self.stacks[y][x] < 100:
-										sameCases.append((y,x))
-									if x_value is None:
-										freeCases.append((y,x))
-							
-							if len(sameCases) > 0:
-								if sameCases[0] == 'hands':
-									self.handsStack += 1
-								else:
-									y,x = sameCases[0]
-									if self.slots[y][x] is not None:
-										self.stacks[y][x] += 1
-								self.player.engine.entitiesManager.killEntity(caseInfo['x_value'])
-
-							elif len(freeCases) > 0:
-								if freeCases[0] == 'hands':
-									self.hands = caseInfo['x_value']
-									self.handsStack = 1
-								else:
-									y,x = freeCases[0]
-									self.slots[y][x] = caseInfo['x_value']
-									self.stacks[y][x] = 1
-								self.player.engine.entitiesManager.killEntity(caseInfo['x_value'])
-	
-						else:
-							self.isOpen = False
-
-class SkillTree():
-	def __init__(self,player):
-		self.player = player
-		self.skills = self.player.data['skillTree']['skills']
-		self.isOpen = False
-		self.font = pygame.font.Font('./assets/fonts/Pixel Digivolve.otf', 20*self.player.engine.param['HudScale'])
-		self.engine = self.player.engine
-	
-	def drawHud(self,passThrough):
-		if passThrough['currentHUD'] not in ['skillTree']:
-			return
-
-		if passThrough['currentHUD'] == 'skillTree':
-			self.skillTree(passThrough)
-			
-	def skillTree(self,passThrough):
-		scale = passThrough['HudScale']
-		width, height = self.engine.HUDSurface.get_size()
-
-		percentageHeight = 1.2
-		percentageWidth = 1.2
-
-		coords = {
-			'left': (width - width/percentageWidth)/2,
-			'top': (height - height/percentageHeight)/2,
-			'height': height/percentageHeight,
-			'width': width/percentageWidth
-		}
-
-		transparency = 150
-		radius = 10
-		radius *= scale
-
-		rects, circles = gFunction.createRectWithRoundCorner(0,0,coords['width'],coords['height'],radius)
-		popup = pygame.Surface((coords['width'],coords['height']), pygame.SRCALPHA)
-		for rect in rects:
-			pygame.draw.rect(popup,(0,0,0,transparency),Rect(rect))
-		for circle in circles:
-			pygame.draw.circle(popup,(0,0,0,transparency),circle,radius)
-
-		self.engine.HUDSurface.blit(popup, (coords['left'], coords['top']))
-		
-		# logic for first part -> skills and skills bar
-		#### logic for left part -> skills name
-
-		heightOfEachPart = (coords['height']/2) / (len(self.skills) + 1)
-		imgWidth, imgHeight = self.font.size('h')
-		marginTop = heightOfEachPart - imgHeight
-
-		img = self.font.render("Skill tree", False, (255, 242, 0))
-		imgShadow = self.font.render("Skill tree", False, (255,200,0))
-		imgLeft, imgTop = coords['left'] + marginTop, coords['top']  + marginTop/1.5
-		self.engine.HUDSurface.blit(imgShadow,(imgLeft+scale, imgTop+scale))
-		self.engine.HUDSurface.blit(img,(imgLeft, imgTop))
-
-		if self.player.stats['points'] > 0:
-			img = self.font.render(f"{self.player.stats['points']} points", False, (255, 242, 0))
-			imgShadow = self.font.render(f"{self.player.stats['points']} points", False, (255,200,0))
-			imgLeft, imgTop = coords['left'] + coords['width'] - img.get_width() - marginTop, coords['top']  + marginTop/1.5
-			self.engine.HUDSurface.blit(imgShadow,(imgLeft+scale, imgTop+scale))
-			self.engine.HUDSurface.blit(img,(imgLeft, imgTop))
-
-		longerSkillName = 0
-		for i in range(len(self.skills)):
-			img = self.font.render(self.skills[i]['displayName'], False, (255, 242, 0))
-			if img.get_width() > longerSkillName:
-				longerSkillName = img.get_width()
-			imgShadow = self.font.render(self.skills[i]['displayName'], False, (255,200,0))
-			imgLeft = coords['left'] + marginTop
-			imgTop = coords['top'] + (i+1)*heightOfEachPart + marginTop/1.5
-			self.engine.HUDSurface.blit(imgShadow,(imgLeft+scale, imgTop+scale))
-			self.engine.HUDSurface.blit(img,(imgLeft, imgTop))
-			
-		#### logic for right part -> skills bars
-
-		barCoord = {}
-		barCoord['btnWidth'] = img.get_height()
-		barCoord['left'] = imgLeft + longerSkillName + marginTop + barCoord['btnWidth'] + marginTop/1.5
-		barCoord['width'] = coords['width'] - barCoord['left'] + coords['left'] - marginTop 
-		barRadius = 5*scale
-
-		for i in range(len(self.skills)):
-			top = (i+1)*heightOfEachPart + coords['top'] + marginTop/1.5
-
-			collisionRect = Rect((barCoord['left'] - barCoord['btnWidth'] - marginTop/1.5 ,top),(int(barCoord['btnWidth']),int(barCoord['btnWidth'])))
-			isHover = collisionRect.collidepoint(passThrough['mousePos'])
-			btnColor = (133, 125, 66) if isHover else (255, 147, 38)
-			btnColor = btnColor if self.player.stats['points'] > 0 and self.skills[i]['steps'] < self.skills[i]['maxSteps'] else (133, 125, 66)
-
-			rects, circles = gFunction.createRectWithRoundCorner(barCoord['left'] - barCoord['btnWidth'] - marginTop/1.5 ,top,int(barCoord['btnWidth']),int(barCoord['btnWidth']),barRadius)
-			for rect in rects:
-				pygame.draw.rect(self.engine.HUDSurface,btnColor,Rect(rect))
-			for circle in circles:
-				pygame.draw.circle(self.engine.HUDSurface,btnColor,circle,barRadius)
-
-			img = self.font.render("+", False, (255, 255, 255))
-			self.engine.HUDSurface.blit(img,(barCoord['left'] - barCoord['btnWidth'] - marginTop/1.5 + img.get_width()/2 + scale/2, top - scale))
-
-			leftOfEachPoly = (barCoord['width'] - barRadius) /self.skills[i]['maxSteps']
-			barStepsSteep = leftOfEachPoly - 2*scale if leftOfEachPoly - 2*scale < 10*scale else 10*scale
-			for step in range(self.skills[i]['maxSteps']):
-				color = (255, 147, 38) if step+1 <= self.skills[i]['steps'] else (133, 125, 66)
-				if step == 0:
-					pygame.draw.circle(self.engine.HUDSurface,color,(barCoord['left'] + barRadius,top + barRadius),barRadius,draw_top_left=True)
-					pygame.draw.circle(self.engine.HUDSurface,color,(barCoord['left'] + barRadius,top + img.get_height() - barRadius),barRadius,draw_bottom_left=True)
-					pygame.draw.rect(self.engine.HUDSurface,color,Rect(barCoord['left'] + step*leftOfEachPoly, top + barRadius, barRadius, img.get_height() - 2*barRadius))
-
-					points = [
-						(barCoord['left'] + step*leftOfEachPoly + barRadius, top),
-						(barCoord['left'] + step*leftOfEachPoly + barStepsSteep + leftOfEachPoly - 2*scale, top),
-						(barCoord['left'] + step*leftOfEachPoly + leftOfEachPoly - 2*scale, top+img.get_height()),
-						(barCoord['left'] + step*leftOfEachPoly + barRadius, top+img.get_height())
-					]
-					#Rect(barCoord['left'] + step*leftOfEachPoly, top, leftOfEachPoly-scale, img.get_height())
-					pygame.draw.polygon(self.engine.HUDSurface, color, points)		
-				elif step == self.skills[i]['maxSteps']-1:
-					pygame.draw.circle(self.engine.HUDSurface,color,(barCoord['left'] + step*leftOfEachPoly + leftOfEachPoly - 2*scale,top + barRadius),barRadius,draw_top_right=True)
-					pygame.draw.circle(self.engine.HUDSurface,color,(barCoord['left'] + step*leftOfEachPoly + leftOfEachPoly - 2*scale,top + img.get_height() - barRadius),barRadius,draw_bottom_right=True)
-					pygame.draw.rect(self.engine.HUDSurface,color,Rect(barCoord['left'] + step*leftOfEachPoly + leftOfEachPoly - 2*scale, top + barRadius, barRadius, img.get_height() - 2*barRadius))
-
-					points = [
-						(barCoord['left'] + step*leftOfEachPoly + barStepsSteep, top),
-						(barCoord['left'] + step*leftOfEachPoly + leftOfEachPoly - 2*scale, top),
-						(barCoord['left'] + step*leftOfEachPoly + leftOfEachPoly - 2*scale, top+img.get_height()),
-						(barCoord['left'] + step*leftOfEachPoly, top+img.get_height())
-					]
-					#Rect(barCoord['left'] + step*leftOfEachPoly, top, leftOfEachPoly-scale, img.get_height())
-					pygame.draw.polygon(self.engine.HUDSurface, color, points)
-				else:
-					points = [
-						(barCoord['left'] + step*leftOfEachPoly + barStepsSteep, top),
-						(barCoord['left'] + step*leftOfEachPoly + barStepsSteep + leftOfEachPoly - 2*scale, top),
-						(barCoord['left'] + step*leftOfEachPoly + leftOfEachPoly - 2*scale, top+img.get_height()),
-						(barCoord['left'] + step*leftOfEachPoly, top+img.get_height())
-					]
-					#Rect(barCoord['left'] + step*leftOfEachPoly, top, leftOfEachPoly-scale, img.get_height())
-					pygame.draw.polygon(self.engine.HUDSurface, color, points)
-
-			# handle clicks requests
-			if passThrough['mousePosClick'] is not None:
-				if collisionRect.collidepoint(passThrough['mousePosClick']):
-					skill = self.skills[i]
-					if self.player.stats['points'] > 0 and skill['steps'] < skill['maxSteps']:
-						skill['steps'] += 1
-						self.player.stats['points'] -= 1
-						self.player.applyNewStats()
-
-		# logic for second part -> global player stats
-
-class Entity(gregngine.Entity):
-	def __init__(self, param):
-		super().__init__(param)
-	
-	def death(self):
-		print('im dead')
-		if self.data['type'] == 'monster':
-			for drop in self.stats['drops']:
-				rng = random.randint(0,100)
-				print(f"{rng} of {drop['rate']} to spawn {drop['loot']}")
-				if drop['rate'] >= rng:
-					item = Item({'itemRepr': drop['loot'], "engine": self.engine,'x': int(self.x),'y': int(self.y),'isGrounded':True})
-					self.engine.entitiesManager.addEntity(item)
-			
-			self.engine.player.addExp(self.stats['exp'])
-
-class Player(gregngine.Entity):
-	def __init__(self, param):
-		super().__init__(param)
-
-		self.speed = self.stats["normalSpeed"]
-		self.running = False
-
-		self.camera = None
-		self.inventory = Inventory(self)
-		self.skillTree = SkillTree(self)
-		self.applyNewStats()
-
-		self.stats["energy"] = self.stats["maxEnergy"]
-		self.stats["level"] = 1
-		self.applyExpFromLevel()
-
-		self.moneyLimit = 9999999
-		self.stats["money"] = 0
-
-		self.lastAttackTime = time.time()
-		self.isAttacking = False
-		self.atkAnim = {'lastStep':-1,'lastTime':0}
-
-		self.animator.setFrame()
-
-	def applyExpFromLevel(self):
-		expTable = self.data['experienceTable']
-		self.stats["experience"] = 0
-		self.stats["nextLevelExp"] = expTable[str(self.stats["level"])]['nextLevelExp']
-
-	def initPostParent(self):
-		self.bigFont = pygame.font.Font('./assets/fonts/Pixel Digivolve.otf', 30*self.engine.param['HudScale'])
-		self.smallFont = pygame.font.Font('./assets/fonts/Pixel Digivolve.otf', 14*self.engine.param['HudScale'])
-
-	def applyNewStats(self):
-		for skill in self.data['skillTree']['skills']:
-			if skill['statPerStep']['mode'] == 'add':
-				self.stats[skill['statToModify']] = self.data['baseStats'][skill['statToModify']] + skill['steps']*skill['statPerStep']['value']
-			elif skill['statPerStep']['mode'] == 'sub':
-				self.stats[skill['statToModify']] = self.data['baseStats'][skill['statToModify']] - skill['steps']*skill['statPerStep']['value']
-
-		self.data['animation']['attack']['stepTime'] = self.stats['atkPerSec']/len(self.data['animation']['attack']['steps'])
-
-	def addMoney(self,money):
-		addition = self.stats["money"] + money
-		if addition > self.moneyLimit:
-			self.stats["money"] = self.moneyLimit
-		else:
-			self.stats["money"] = addition
-	
-	def addExp(self,exp):
-		if self.stats["level"] < 100:
-			addition = self.stats["experience"] + exp
-			if addition > self.stats["nextLevelExp"]:
-				self.stats["level"] += 1
-				self.stats["points"] += 1
-				rest = addition - self.stats["nextLevelExp"]
-				self.applyExpFromLevel()
-				self.addExp(rest)
-			else:
-				self.stats["experience"] = addition
-
-	def healOf(self,ammount):
-		self.stats['health'] += ammount
-		if self.stats['health'] > self.stats['maxHealth']:
-			self.stats['health'] = self.stats['maxHealth']
-
-	def drawHud(self, xPos, yPos, passThrough):
-		self.inventory.drawHud(passThrough)
-		self.skillTree.drawHud(passThrough)
-
-		if passThrough['currentHUD'] == 'main':
-			scale = passThrough['HudScale']
-
-			top = 8
-			left = 8
-
-			expValue = 143*(self.stats['experience']/self.stats['nextLevelExp']) 
-			expValue = expValue if expValue > 7 else 7
-
-			healValue = 94*(self.stats['health']/self.stats['maxHealth']) 
-			healValue = healValue if healValue > 7 else 7
-
-			energyValue = 94*(self.stats['energy']/self.stats['maxEnergy']) 
-			energyValue = energyValue if energyValue > 7 else 7
-
-			rectCoords = [
-				{ # first rect
-					'top': 0,
-					'left': 0,
-					'width': 50,
-					'height': 50
-				},{ # ---- exp bar ------
-					# second bottom rect 
-					'top': 39,
-					'left': 0,
-					'width': 150,
-					'height': 11
-				},{ # exp bar behind rect
-					'top': 43,
-					'left': 2,
-					'width': 143,
-					'height': 3,
-					'color': (85,85,85,255)
-				},{ # exp bar front rect
-					'top': 41,
-					'left': 2,
-					'width': expValue,
-					'height': 7,
-					'color': (0, 135, 255,255)
-				},{ # ---- heal bar ----
-					# heal bar outline bottom rect
-					'top': 2,
-					'left': 52,
-					'width': 98,
-					'height': 11
-				},{ # heal bar behind rect
-					'top': 6,
-					'left': 55,
-					'width': 91,
-					'height': 3,
-					'color': (85,85,85,255)
-				},{ # heal bar front rect
-					'top': 4,
-					'left': 54,
-					'width': healValue,
-					'height': 7,
-					'color': (255,80,0,255)
-				},{ # ---- energy bar ----
-					# energy bar outline bottom rect
-					'top': 15,
-					'left': 52,
-					'width': 98,
-					'height': 11
-				},{ # energy bar behind rect
-					'top': 19,
-					'left': 55,
-					'width': 91,
-					'height': 3,
-					'color': (85,85,85,255)
-				},{ # energy bar front rect
-					'top': 17,
-					'left': 54,
-					'width': energyValue,
-					'height': 7,
-					'color': (0,80,255,255)
-				}
-			]
-			transparency = 225
-			Radius = 10
-
-			case = pygame.Surface((150*scale,50*scale), pygame.SRCALPHA)
-			for coords in rectCoords:
-				radius = Radius*scale if Radius*scale <= coords['height']*scale/2 else coords['height']*scale/2
-				color = coords['color'] if 'color' in coords else (0,0,0,transparency)
-				rects, circles = gFunction.createRectWithRoundCorner(coords['left']*scale,coords['top']*scale,coords['width']*scale,coords['height']*scale,radius)
-				for rect in rects:
-					pygame.draw.rect(case,color,Rect(rect))
-				for circle in circles:
-					pygame.draw.circle(case,color,circle,radius)
-				
-				if coords == rectCoords[0]:
-					pygame.draw.rect(case,(0,0,0,transparency),
-						Rect(rectCoords[0]['width']*scale,(rectCoords[1]['top']-Radius)*scale,Radius*scale,Radius*scale)) #little concave circle
-					pygame.draw.circle(case,(0,0,0,0),((rectCoords[0]['width']+Radius)*scale,(rectCoords[1]['top']-Radius)*scale),Radius*scale,draw_bottom_left=True) #little concave circle mask
-			
-			self.engine.HUDSurface.blit(case, (left*scale, top*scale))
-
-			# render level text
-			img = self.bigFont.render(str(self.stats['level']), False, (255, 242, 0))
-			imgLeft = int(left*scale + rectCoords[0]['width']*scale/2 - img.get_width()/2) + scale
-			imgTop = int(top*scale + rectCoords[1]['top']*scale/2 - img.get_height()/2) + scale
-			self.engine.HUDSurface.blit(img,(imgLeft, imgTop))
-
-			# render money text
-			img = self.smallFont.render(str(self.stats['money']) + ' $', False, (255, 242, 0))
-			imgShadow = self.smallFont.render(str(self.stats['money']) + ' $', False, (0, 0, 0, 0.5))
-			imgLeft = int(left*scale + rectCoords[1]['width']*scale - img.get_width()) + scale
-			imgTop = int(top*scale + rectCoords[7]['top']*scale + img.get_height()/2 )
-			self.engine.HUDSurface.blit(imgShadow,(imgLeft+scale, imgTop+scale))
-			self.engine.HUDSurface.blit(img,(imgLeft, imgTop))
-
-	def drawHudBar(self,value,max,color,passThrough,coords):
-		"""
-		coords = {
-			'top',
-			'left',
-			'width',
-			'height',
-			'outline'
-		}
-		"""
-		scale = passThrough['HudScale']
-		x = ((value * 100) / max) /100 if value > 0 else 0
-
-		rect = gFunction.createRectOutlined(x, coords['left']*scale, coords['top']*scale, coords['width']*scale, coords['height']*scale, coords['outline']*scale)
-		pygame.draw.rect(self.engine.HUDSurface,(0,0,0),rect[0])
-
-		if x > 0:
-			pygame.draw.rect(self.engine.HUDSurface,color,rect[1])
-
-	def move(self,dTime):
-		x,y = self.velocity['x'],self.velocity['y']
-
-		if self.running and self.stats['energy']  > 0 and (x != 0 or y !=0):
-			x,y = x*self.stats["maxSpeed"],y*self.stats["maxSpeed"]
-			if self.stats['energy']  < 1:
-				self.stats['energy'] -= 10
-			else:
-				self.stats['energy'] -= 30*dTime*0.001
-		else:
-			x,y = x*self.stats["normalSpeed"],y*self.stats["normalSpeed"]
-
-		x,y = gFunction.normalize(x * dTime * 0.001,y * dTime * 0.001)
-		x,y = round(x,3), round(y,3)
-
-		self.x += x
-		self.y += y
-
-		camSmoothOffset = self.data['camera']['smoothOffset']
-
-		if self.x - self.camera.x > self.param['x']+camSmoothOffset or self.x - self.camera.x < self.param['x']-camSmoothOffset:
-			self.camera.move(x,0)
-		if self.y - self.camera.y > self.param['y']+camSmoothOffset or self.y - self.camera.y < self.param['y']-camSmoothOffset:
-			self.camera.move(0,y)
-
-		self.setVelocity(0,0)
-
-	def increaseEnergy(self,dtime):
-		if self.stats['energy'] < self.stats['maxEnergy']:
-			self.stats['energy'] += 10*dtime*0.001
-			if self.stats['maxEnergy'] < self.stats['energy']:
-				self.stats['energy'] = self.stats['maxEnergy']
-
-	def setCameraOffset(self,x,y):
-		self.camera.xOffset = -(x/2) + self.param['x']
-		self.camera.yOffset = -(y/2) + self.param['y']
-
-	def attack(self):
-		if self.inventory.hands is None:
-			atkPerSec = self.stats['atkPerSec']
-			damage = self.stats['atkBaseDamage']
-		elif self.inventory.hands.data['itemType'] == 'consumable':
-			atkPerSec = self.stats['atkPerSec']
-		else:
-			atkPerSec = self.stats['atkPerSec']*weaponTypeAPSMultiplier.get(self.inventory.hands.data['weaponType'])
-			damage = self.stats['atkBaseDamage'] + self.inventory.hands.data['stats']['damage']
-
-		currentTime = time.time()   
-		if currentTime-self.lastAttackTime >= atkPerSec and not self.isAttacking:
-			self.lastAttackTime = currentTime 
-
-			if self.inventory.hands is not None and self.inventory.hands.data['itemType'] == 'consumable':
-				self.consumeItemInHand()
-				return None
-
-			dico = {
-				'TOP WALK': 'top',
-				'BOTTOM WALK': 'bottom',
-				'RIGHT WALK': 'right',
-				'LEFT WALK': 'left',
-			}
-
-			anim = self.animator.animation if self.animator.animation != 'IDLE' else self.animator.lastAnimation
-			lookingTo = dico.get(anim)
-			self.atkLookingTo = lookingTo
-
-			attackRect = self.rect
-
-			if lookingTo == 'left':
-				attackRect.left -= self.param['pixelSize']
-			elif lookingTo == 'right':
-				attackRect.left += self.param['pixelSize']
-			elif lookingTo == 'top':
-				attackRect.top -= self.param['pixelSize']
-			elif lookingTo == 'bottom':
-				attackRect.top += self.param['pixelSize']
-
-			self.isAttacking = True
-
-			return {'entity':self,'attackRect':attackRect,'damage':damage}
-
-		else:
-			return None
-
-	def attackAnimation(self):
-		if self.isAttacking:
-			t = self.atkAnim['lastTime'] + self.data['animation']['attack']['stepTime']
-			if t < time.time():
-				self.atkAnim['lastTime'] = time.time()
-				self.atkAnim['lastStep'] += 1
-
-			if self.atkAnim['lastStep'] == len(self.data['animation']['attack']['steps']):
-				self.isAttacking = False
-				self.atkAnim['lastStep'] = -1
-				self.atkAnim['lastTime'] = 0
-
-			elif self.inventory.hands is not None:
-				currentStep = self.data['animation']['attack']['steps'][self.atkAnim['lastStep']]
-				animationSurface = pygame.Surface((int(3*self.param['pixelSize']),int(3*self.param['pixelSize'])), pygame.SRCALPHA)
-
-				sprite = self.inventory.hands.sprite
-				tilemap = self.inventory.hands.data['tilemap']
-
-				itemScale = (int(self.param['pixelSize']*tilemap['scaling']/tilemap['ratio']),int(self.param['pixelSize']*tilemap['scaling']))
-				sprite = pygame.transform.scale(sprite.convert_alpha(),(itemScale))
-				sprite = pygame.transform.rotate(sprite,currentStep['rotation'])
-
-				animationSurface.blit(sprite,((currentStep['x']/50) * self.param['pixelSize'],(currentStep['y']/50) * self.param['pixelSize']))
-				if self.atkLookingTo == 'top':
-					animationSurface = pygame.transform.rotate(animationSurface,90)
-				elif self.atkLookingTo == 'bottom':
-					animationSurface = pygame.transform.rotate(animationSurface,-90)
-				elif self.atkLookingTo == 'left':
-					animationSurface = pygame.transform.flip(animationSurface,True,False)
-
-				return animationSurface
-
-		return None
-
-	def consumeItemInHand(self):
-		if self.inventory.hands.data["itemType"] == 'consumable':
-			if 'heal' in self.inventory.hands.data['stats']:
-				self.healOf(self.inventory.hands.data['stats']['heal'])
-
-			if self.inventory.handsStack > 1:
-				self.inventory.handsStack -= 1
-			else:
-				self.inventory.hands = None
-				self.inventory.handsStack = 0
-			
-	def draw(self, xStart, yStart, passThrough):
-		super().draw(xStart, yStart, passThrough)
+from statics import *
+
+#player scirpts import
+from scripts.player.player import *
+from scripts.player.inventory import *
+from scripts.player.skillTree import *
+
+# other scripts import
+from scripts.HUDMenuManager import *
+from scripts.item import *
+from scripts.entity import *
 
 class Engine(gregngine.Engine):
 	def __init__(self, param):
@@ -982,12 +45,6 @@ class Engine(gregngine.Engine):
 		self.collisions = {}
 
 		self.damagesInfos = []
-
-		#self.rezizeSprites()
-
-	def rezizeSprites(self):
-		for sprite in self.world.sprites:
-			self.world.sprites[sprite] = pygame.transform.scale(self.world.sprites[sprite].convert_alpha(),(self.newPixelScale,self.newPixelScale))
 
 	def loadScreenSaves(self):
 		super().loadSaves()
@@ -1062,6 +119,52 @@ class Engine(gregngine.Engine):
 
 		print('data loaded')
 		return True
+
+	def saveGame(self):
+		with open(self.savePath + '/data.save', 'wb') as f:
+			data = {}
+			
+			data['playerPos'] = (self.player.x,self.player.y)
+			data['playerOrientation'] = self.player.animator.lastAnimation
+
+			data['playerInv'] = {}
+			data['playerInv']['hands'] = None if self.player.inventory.hands is None else self.player.inventory.hands.data['name']
+			data['playerInv']['handsStack'] = self.player.inventory.handsStack
+			data['playerInv']['slots'] = []
+			for row in self.player.inventory.slots:
+				r = []
+				for col in row:
+					if col is not None:
+						r.append(col.data['name'])
+					else:
+						r.append(None)
+				data['playerInv']['slots'].append(r)
+
+			data['playerInv']['stacks'] = self.player.inventory.stacks
+			data['playerData'] = self.player.data
+			data['playerData']['skillTree']['skills'] = self.player.skillTree.skills
+
+			data['entities'] = []
+			for entitie in self.entitiesManager.entities:
+				dico = {}
+				if 'entityRepr' in entitie.param and entitie.param['entityRepr'] == 'player':
+					continue 
+
+				if 'entityRepr' in entitie.param and entitie.param['entityRepr'] != 'player':
+					dico['type'] = 'entitie'
+					dico['repr'] = entitie.param['entityRepr']
+					dico['stats'] = entitie.stats
+
+				elif 'itemRepr' in entitie.param:
+					dico['type'] = 'item'
+					dico['repr'] = entitie.param['itemRepr']
+
+				dico['x'], dico['y'] = entitie.x, entitie.y
+				data['entities'].append(dico)
+
+			pickle.dump(data, f, pickle.HIGHEST_PROTOCOL)
+
+		print('Save')
 
 	def loadGame(self):
 		#savesLoaded = self.loadSaves()
